+++ services/authentication/app/src/main.rs
@@ -15,42 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use anyhow::{Context, Result};
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-use std::thread;
-use teaclave_service_app_utils::{register_signals, TeaclaveServiceLauncher};
+use anyhow::Result;
+use teaclave_service_app_utils::launch_teaclave_service;
 
 const PACKAGE_NAME: &str = env!("CARGO_PKG_NAME");
 
 fn main() -> Result<()> {
-    env_logger::init_from_env(
-        env_logger::Env::new()
-            .filter_or("TEACLAVE_LOG", "RUST_LOG")
-            .write_style_or("TEACLAVE_LOG_STYLE", "RUST_LOG_STYLE"),
-    );
-
-    let launcher = Arc::new(TeaclaveServiceLauncher::new(
-        PACKAGE_NAME,
-        "runtime.config.toml",
-    )?);
-    let launcher_ref = launcher.clone();
-    thread::spawn(move || {
-        let _ = launcher_ref.start();
-        unsafe { libc::raise(signal_hook::SIGTERM) }
-    });
-
-    let term = Arc::new(AtomicBool::new(false));
-    register_signals(term.clone()).context("Failed to register signal handler")?;
-
-    while !term.load(Ordering::Relaxed) {
-        thread::park();
-    }
-
-    launcher.finalize();
-    unsafe {
-        launcher.destroy(); // force to destroy the enclave
-    }
-
-    Ok(())
+    launch_teaclave_service(PACKAGE_NAME)
 }
